requirements.txt
Flask==3.1.0
mysql-connector-python==9.2.0
pywin32==311
requests==2.32.3
cryptography==46.0.5
